+++ packages/controller-utils/src/index.test.ts
@@ -5,6 +5,7 @@ describe('@metamask/controller-utils', () => {
     expect(Object.keys(allExports)).toMatchInlineSnapshot(`
       [
         "encodeFunctionData",
+        "decodeFunctionResult",
         "BrokenCircuitError",
         "CircuitState",
         "CockatielEventEmitter",